feat: add custom name option for demo creation
The ability to assign custom names during the creation of demos has been
implemented. This flexibility allows easier identification and
management of demos. The default demo name generation using ulid() is
still retained as a fallback for when no custom name is given, ensuring
backward compatibility. This feature delivers an enhanced user
experience as well as improved administrative management of demos.

Please note: the new `--name` or `-n` CLI argument can be used to
specify a custom name.