feat: add MIT license
A MIT license file has been introduced to our project. This license
offers valuable legal protection, granting permissions for free usage,
modification, and distribution of the software and its documentation.
This is a standard requirement for open-source software and also
outlines the lack of warranty and states the conditions for using the
software.

This step further formalizes our project and promotes collaboration and
contributions by clearly stating that users are free to use, modify, and
distribute the code.