build: disable windows binary publishing workflow
The changes in this commit temporarily suspend the GitHub Actions
workflow responsible for building and shipping Windows-compatible
binaries on each release. This is suggested by the commenting out of the
'publish_win_bin' job within the 'release-please.yml' workflow.

Although this workflow has previously been necessary for our release
pipeline, current conditions necessitate its disablement for an unknown
period. It's worthwhile noting that this doesn't remove or alter the job
but merely disables it, leaving it in a state where it could be
reactivated as effortlessly as by uncommenting the pertaining sections.

By optimistically assuming that this is a short-term change, any
negative impact on our build pipeline and release process should be
quite minimal and quickly reversible.

Refer back to this commit when the alterations need to be reverted, or
for any further investigatory needs.