build: switch to oven-sh/setup-bun action
This commit replaces the use of antongolub/action-setup-bun action with
oven-sh/setup-bun across the all workflows in our GitHub
release-please.yml file. The change applies not only to the executed
steps but also to the commented-out examples, ensuring consistency and
avoiding future confusion.

The oven-sh/setup-bun action is now preferred because provides
additional functionality and improvements over the previously used
action. It enhances our build and release processes, ensuring smoother
operations and more efficient workflows.

This change doesn't affect the way our workflows run or impact
user-facing features of the project. It is an internal enhancement that
brings better maintainability and scalability to our CI/CD
infrastructure.