feat: Improve workspace creation and template handling
This update introduces enhancements to the creation of workspaces and
management of templates. It allows users to choose whether or not
immediately open an editor upon workspace creation by leveraging a new
--no-open flag. The caching mechanism in multi-fetch is also improved by
making a clone of the response.

Further, changes are made to better handle how template files are
created and written. A new .demorc file is introduced to manage scripts
including install scripts. Execution of commands now inherit stdio to
allow for interactive scripts.

This commit also introduces new scripts for tarball creation and adjusts
handling of default templates. These modifications will result in more
flexibility and improved user experience when using the tools to create
new workspaces.