feat: Improve template generation and caching mechanism
This commit introduces several enhancements that optimize the template
generation and caching process. Firstly, a cacheRequests option was
added to allow users to customize the location of cache files. This
brought changes to the loadLocalConfig function and made necessary
modifications to the multi_fetch module defaultCacheFolder to correspond
with the globalConfig's cache location.

Secondly, the create_template module was updated to improve the handling
of template creation and writing. A temporary location is now used
during the process, and this location can be kept for debugging if
required. The tar extraction method was also changed from the tar-stream
package to the tar command line utility for better performance and to
resolve compatibility issues. As part of these modifications, the
Template type is now an AsyncIterable, ensuring that the template
generation process can handle asynchronous operations more effectively.

Finally, updates were made to address fetching of default templates, now
directly fetched from the repository's raw content URL. This provides
more accuracy and improved reliability when fetching templates.

These changes seek to enhance performance, provide more flexibility and
improve reliability in the overall template creation and caching
process.